+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Only partial changelog, [commit history](https://framagit.org/SondagePro-LimeSur
 
 - Batch size by surveys
 - Different batch size for reminder and invitation by surveys
-- Simulate system (need update code) for purpose testing
+- Allow debug and simulate system in command
 
 ### Fix